+++ frontend/openchat-agent/src/services/community/community.client.ts
@@ -76,6 +76,7 @@ import {
     addBotResponse,
     updateBotResponse,
     removeBotResponse,
+    bigintTo32bit,
 } from "../common/chatMappersV2";
 import type {
     AddMembersToChannelResponse,
@@ -316,7 +317,7 @@ export class CommunityClient extends CandidService {
         return this.executeMsgpackUpdate(
             "claim_prize",
             {
-                channel_id: Number(channelId),
+                channel_id: bigintTo32bit(channelId),
                 message_id: messageId,
             },
             claimPrizeResponse,
@@ -334,7 +335,7 @@ export class CommunityClient extends CandidService {
         return this.executeMsgpackUpdate(
             "add_members_to_channel",
             {
-                channel_id: Number(chatId.channelId),
+                channel_id: bigintTo32bit(chatId.channelId),
                 user_ids: userIds.map(principalStringToBytes),
                 added_by_name: username,
                 added_by_display_name: displayName,
@@ -357,7 +358,7 @@ export class CommunityClient extends CandidService {
         return this.executeMsgpackUpdate(
             "add_reaction",
             {
-                channel_id: Number(chatId.channelId),
+                channel_id: bigintTo32bit(chatId.channelId),
                 username,
                 display_name: displayName,
                 message_id: messageId,
@@ -391,7 +392,7 @@ export class CommunityClient extends CandidService {
         return this.executeMsgpackUpdate(
             "change_channel_role",
             {
-                channel_id: Number(chatId.channelId),
+                channel_id: bigintTo32bit(chatId.channelId),
                 user_id: principalStringToBytes(userId),
                 new_role: apiMemberRole(newRole),
             },
@@ -446,7 +447,7 @@ export class CommunityClient extends CandidService {
         return this.executeMsgpackUpdate(
             "decline_invitation",
             {
-                channel_id: Number(chatId.channelId),
+                channel_id: bigintTo32bit(chatId.channelId),
             },
             declineInvitationResponse,
             CommunityDeclineInvitationArgs,
@@ -458,7 +459,7 @@ export class CommunityClient extends CandidService {
         return this.executeMsgpackUpdate(
             "delete_channel",
             {
-                channel_id: Number(chatId.channelId),
+                channel_id: bigintTo32bit(chatId.channelId),
             },
             deleteGroupResponse,
             CommunityDeleteChannelArgs,
@@ -474,7 +475,7 @@ export class CommunityClient extends CandidService {
         return this.executeMsgpackQuery(
             "deleted_message",
             {
-                channel_id: Number(chatId.channelId),
+                channel_id: bigintTo32bit(chatId.channelId),
                 message_id: messageId,
                 thread_root_message_index: threadRootMessageIndex,
             },
@@ -494,7 +495,7 @@ export class CommunityClient extends CandidService {
         return this.executeMsgpackUpdate(
             "delete_messages",
             {
-                channel_id: Number(chatId.channelId),
+                channel_id: bigintTo32bit(chatId.channelId),
                 message_ids: messageIds,
                 as_platform_moderator: asPlatformModerator,
                 thread_root_message_index: threadRootMessageIndex,
@@ -529,7 +530,7 @@ export class CommunityClient extends CandidService {
                 return this.executeMsgpackUpdate(
                     "edit_message",
                     {
-                        channel_id: Number(chatId.channelId),
+                        channel_id: bigintTo32bit(chatId.channelId),
                         thread_root_message_index: threadRootMessageIndex,
                         content: apiMessageContent(content ?? message.content),
                         message_id: message.messageId,
@@ -625,7 +626,7 @@ export class CommunityClient extends CandidService {
         maxEvents: number = MAX_EVENTS,
     ): Promise<EventsResponse<ChatEvent>> {
         const args = {
-            channel_id: Number(chatId.channelId),
+            channel_id: bigintTo32bit(chatId.channelId),
             thread_root_message_index: threadRootMessageIndex,
             max_messages: MAX_MESSAGES,
             max_events: maxEvents,
@@ -675,7 +676,7 @@ export class CommunityClient extends CandidService {
         latestKnownUpdate: bigint | undefined,
     ): Promise<EventsResponse<ChatEvent>> {
         const args = {
-            channel_id: Number(chatId.channelId),
+            channel_id: bigintTo32bit(chatId.channelId),
             thread_root_message_index: threadRootMessageIndex,
             events: eventIndexes,
             latest_known_update: latestKnownUpdate,
@@ -769,7 +770,7 @@ export class CommunityClient extends CandidService {
         maxEvents: number = MAX_EVENTS,
     ): Promise<EventsResponse<ChatEvent>> {
         const args = {
-            channel_id: Number(chatId.channelId),
+            channel_id: bigintTo32bit(chatId.channelId),
             thread_root_message_index: threadRootMessageIndex,
             max_messages: MAX_MESSAGES,
             max_events: maxEvents,
@@ -824,7 +825,7 @@ export class CommunityClient extends CandidService {
         latestKnownUpdate: bigint | undefined,
     ): Promise<EventsResponse<Message>> {
         const args = {
-            channel_id: Number(chatId.channelId),
+            channel_id: bigintTo32bit(chatId.channelId),
             thread_root_message_index: undefined,
             messages: messageIndexes,
             invite_code: mapOptional(this.inviteCode, textToCode),
@@ -891,7 +892,7 @@ export class CommunityClient extends CandidService {
         return this.executeMsgpackUpdate(
             "leave_channel",
             {
-                channel_id: Number(chatId.channelId),
+                channel_id: bigintTo32bit(chatId.channelId),
             },
             leaveGroupResponse,
             CommunityLeaveChannelArgs,
@@ -913,7 +914,7 @@ export class CommunityClient extends CandidService {
         return this.executeMsgpackUpdate(
             "unpin_message",
             {
-                channel_id: Number(chatId.channelId),
+                channel_id: bigintTo32bit(chatId.channelId),
                 message_index: messageIndex,
             },
             unpinMessageResponse,
@@ -926,7 +927,7 @@ export class CommunityClient extends CandidService {
         return this.executeMsgpackUpdate(
             "pin_message",
             {
-                channel_id: Number(chatId.channelId),
+                channel_id: bigintTo32bit(chatId.channelId),
                 message_index: messageIndex,
             },
             pinMessageResponse,
@@ -954,7 +955,7 @@ export class CommunityClient extends CandidService {
         return this.executeMsgpackUpdate(
             "remove_member_from_channel",
             {
-                channel_id: Number(chatId.channelId),
+                channel_id: bigintTo32bit(chatId.channelId),
                 user_id: principalStringToBytes(userId),
             },
             removeMemberFromChannelResponse,
@@ -972,7 +973,7 @@ export class CommunityClient extends CandidService {
         return this.executeMsgpackUpdate(
             "remove_reaction",
             {
-                channel_id: Number(chatId.channelId),
+                channel_id: bigintTo32bit(chatId.channelId),
                 message_id: messageId,
                 reaction,
                 thread_root_message_index: threadRootMessageIndex,
@@ -1002,7 +1003,7 @@ export class CommunityClient extends CandidService {
         return this.executeMsgpackQuery(
             "search_channel",
             {
-                channel_id: Number(chatId.channelId),
+                channel_id: bigintTo32bit(chatId.channelId),
                 max_results: maxResults,
                 users: users.map(principalStringToBytes),
                 search_term: searchTerm,
@@ -1111,7 +1112,7 @@ export class CommunityClient extends CandidService {
         return this.executeMsgpackQuery(
             "selected_channel_initial",
             {
-                channel_id: Number(chatId.channelId),
+                channel_id: bigintTo32bit(chatId.channelId),
             },
             groupDetailsResponse,
             CommunitySelectedChannelInitialArgs,
@@ -1138,7 +1139,7 @@ export class CommunityClient extends CandidService {
         const updatesResponse = await this.executeMsgpackQuery(
             "selected_channel_updates_v2",
             {
-                channel_id: Number(chatId.channelId),
+                channel_id: bigintTo32bit(chatId.channelId),
                 updates_since: previous.timestamp,
             },
             groupDetailsUpdatesResponse,
@@ -1185,7 +1186,7 @@ export class CommunityClient extends CandidService {
             const newEvent =
                 content !== undefined ? { ...event, event: { ...event.event, content } } : event;
             const args = {
-                channel_id: Number(chatId.channelId),
+                channel_id: bigintTo32bit(chatId.channelId),
                 content: apiMessageContent(newEvent.event.content),
                 message_id: newEvent.event.messageId,
                 sender_name: senderName,
@@ -1238,7 +1239,7 @@ export class CommunityClient extends CandidService {
         return this.executeMsgpackUpdate(
             "register_poll_vote",
             {
-                channel_id: Number(chatId.channelId),
+                channel_id: bigintTo32bit(chatId.channelId),
                 thread_root_message_index: threadRootMessageIndex,
                 poll_option: answerIdx,
                 operation: voteType === "register" ? "RegisterVote" : "DeleteVote",
@@ -1255,7 +1256,7 @@ export class CommunityClient extends CandidService {
         return this.executeMsgpackQuery(
             "channel_summary",
             {
-                channel_id: Number(chatId.channelId),
+                channel_id: bigintTo32bit(chatId.channelId),
                 invite_code: mapOptional(this.inviteCode, textToCode),
             },
             (resp) => communityChannelSummaryResponse(resp, this.communityId),
@@ -1333,7 +1334,7 @@ export class CommunityClient extends CandidService {
         return this.executeMsgpackUpdate(
             "toggle_mute_notifications",
             {
-                channel_id: chatId ? Number(chatId.channelId) : undefined,
+                channel_id: chatId ? bigintTo32bit(chatId.channelId) : undefined,
                 mute,
             },
             toggleNotificationsResponse,
@@ -1362,7 +1363,7 @@ export class CommunityClient extends CandidService {
         return this.executeMsgpackUpdate(
             "undelete_messages",
             {
-                channel_id: Number(chatId.channelId),
+                channel_id: bigintTo32bit(chatId.channelId),
                 thread_root_message_index: threadRootMessageIndex,
                 message_ids: [messageId],
             },
@@ -1380,7 +1381,7 @@ export class CommunityClient extends CandidService {
         return this.executeMsgpackQuery(
             "thread_previews",
             {
-                channel_id: Number(chatId.channelId),
+                channel_id: bigintTo32bit(chatId.channelId),
                 threads: threadRootMessageIndexes,
                 latest_client_thread_update: latestClientThreadUpdate,
             },
@@ -1398,7 +1399,7 @@ export class CommunityClient extends CandidService {
         return this.executeMsgpackUpdate(
             "register_proposal_vote",
             {
-                channel_id: Number(channelId),
+                channel_id: bigintTo32bit(channelId),
                 adopt,
                 message_index: messageIdx,
             },
@@ -1424,7 +1425,7 @@ export class CommunityClient extends CandidService {
         return this.executeMsgpackUpdate(
             "update_channel",
             {
-                channel_id: Number(chatId.channelId),
+                channel_id: bigintTo32bit(chatId.channelId),
                 name: name,
                 description,
                 external_url: externalUrl === undefined ? "NoChange" : { SetToSome: externalUrl },
@@ -1581,7 +1582,7 @@ export class CommunityClient extends CandidService {
         newAchievement: boolean,
     ): Promise<FollowThreadResponse> {
         const args = {
-            channel_id: Number(channelId),
+            channel_id: bigintTo32bit(channelId),
             thread_root_message_index: threadRootMessageIndex,
             new_achievement: newAchievement,
         };
@@ -1603,7 +1604,7 @@ export class CommunityClient extends CandidService {
         return this.executeMsgpackUpdate(
             "report_message",
             {
-                channel_id: Number(channelId),
+                channel_id: bigintTo32bit(channelId),
                 thread_root_message_index: threadRootMessageIndex,
                 message_id: messageId,
                 delete: deleteMessage,
@@ -1624,7 +1625,7 @@ export class CommunityClient extends CandidService {
         return this.executeMsgpackUpdate(
             "accept_p2p_swap",
             {
-                channel_id: Number(channelId),
+                channel_id: bigintTo32bit(channelId),
                 thread_root_message_index: threadRootMessageIndex,
                 message_id: messageId,
                 pin,
@@ -1644,7 +1645,7 @@ export class CommunityClient extends CandidService {
         return this.executeMsgpackUpdate(
             "cancel_p2p_swap",
             {
-                channel_id: Number(channelId),
+                channel_id: bigintTo32bit(channelId),
                 thread_root_message_index: threadRootMessageIndex,
                 message_id: messageId,
             },
@@ -1663,7 +1664,7 @@ export class CommunityClient extends CandidService {
             "join_video_call",
             {
                 message_id: messageId,
-                channel_id: Number(channelId),
+                channel_id: bigintTo32bit(channelId),
                 new_achievement: newAchievement,
             },
             joinVideoCallResponse,
@@ -1681,7 +1682,7 @@ export class CommunityClient extends CandidService {
         return this.executeMsgpackUpdate(
             "set_video_call_presence",
             {
-                channel_id: Number(channelId),
+                channel_id: bigintTo32bit(channelId),
                 message_id: messageId,
                 presence: apiVideoCallPresence(presence),
                 new_achievement: newAchievement,
@@ -1700,7 +1701,7 @@ export class CommunityClient extends CandidService {
         return this.executeMsgpackQuery(
             "video_call_participants",
             {
-                channel_id: Number(channelId),
+                channel_id: bigintTo32bit(channelId),
                 message_id: messageId,
                 updated_since: updatesSince,
             },
@@ -1714,7 +1715,7 @@ export class CommunityClient extends CandidService {
         return this.executeMsgpackUpdate(
             "cancel_invites",
             {
-                channel_id: mapOptional(channelId, (cid) => Number(cid)),
+                channel_id: mapOptional(channelId, (cid) => bigintTo32bit(cid)),
                 user_ids: userIds.map(principalStringToBytes),
             },
             (value) => value === "Success",